GEODE-6425: fix rollover of put and update stats
Change put, update, putTimeout, putFailure stats from ints to longs so
that they rollover less often

+++ geode-core/src/integrationTest/java/org/apache/geode/TXJUnitTest.java
@@ -4833,7 +4833,7 @@ public class TXJUnitTest {
     // quick sanity check to make sure perf stats work non-tx
     int creates;
     int destroys;
-    int puts;
+    long puts;
     int invalidates;
 
     creates = cacheStats.getCreates();
