+++ geode-core/src/main/java/org/apache/geode/management/internal/beans/CacheServerBridge.java
@@ -529,7 +529,7 @@ public class CacheServerBridge extends ServerBridge {
       status.setNumOfCacheListenerCalls(stats.getNumOfCacheListenerCalls());
       status.setNumOfGets((int) stats.getNumOfGets());
       status.setNumOfMisses((int) stats.getNumOfMisses());
-      status.setNumOfPuts(stats.getNumOfPuts());
+      status.setNumOfPuts((int) stats.getNumOfPuts());
       status.setNumOfThreads(stats.getNumOfThreads());
       status.setProcessCpuTime(stats.getProcessCpuTime());
       status.setPoolStats(stats.getPoolStats());
